+++ src/Enums/PositionCloseAttribution.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Kraite\Core\Enums;
+
+enum PositionCloseAttribution
+{
+    case External;
+    case Kraite;
+    case Forced;
+    case Unknown;
+}